Add support for Hawakening master query through separate protocol
Protocol 'hawakeningmaster' provides full list of processed server info
