+++ protocols/index.js
@@ -21,6 +21,7 @@ import geneshift from './geneshift.js'
 import goldsrc from './goldsrc.js'
 import gtasao from './gtasao.js'
 import hawakening from './hawakening.js'
+import hawakeningmaster from './hawakeningmaster.js'
 import hexen2 from './hexen2.js'
 import jc2mp from './jc2mp.js'
 import kspdmp from './kspdmp.js'
@@ -71,7 +72,7 @@ import vintagestory from './vintagestory.js'
 
 export {
   armagetron, ase, asa, assettocorsa, battlefield, buildandshoot, cs2d, discord, doom3, eco, epic, factorio, farmingsimulator, ffow,
-  fivem, gamespy1, gamespy2, gamespy3, geneshift, goldsrc, gtasao, hawakening, hexen2, jc2mp, kspdmp, mafia2mp, mafia2online, minecraft,
+  fivem, gamespy1, gamespy2, gamespy3, geneshift, goldsrc, gtasao, hawakening, hawakeningmaster, hexen2, jc2mp, kspdmp, mafia2mp, mafia2online, minecraft,
   minecraftbedrock, minecraftvanilla, minetest, mumble, mumbleping, nadeo, openttd, palworld, quake1, quake2, quake3, renegadex, renegadexmaster, rfactor, ragemp, samp,
   satisfactory, soldat, savage2, starmade, starsiege, teamspeak2, teamspeak3, terraria, toxikk, tribes1, tribes1master, unreal2, ut3, valve,
   vcmp, ventrilo, warsow, eldewrito, beammpmaster, beammp, dayz, theisleevrima, xonotic, altvmp, vintagestorymaster, vintagestory
